Implement constraints using `Comparable` interface
Allow building the constraints

- `maximum` and `exclusiveMaximum`,
- `minimum` and `exclusiveMinimum`

for any type `T`, using a `Comparable<T>`.

Previously, these constraints were only available for subtypes of
`Number` and relied on converting to `Double` for comparing numbers of
different types. This could cause issues when comparing values which
cannot be cleanly converted from the original number type.

For compatibility and convenience, we provide overloads for these
constraints that accept an `Int` as the limit value as this seems like
a rather common case.

Closes #180